requirements.txt
python-binance
ta-lib
websockets
requests
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
numpy==2.1.2
ta==0.11.0
groq==0.11.0
websockets==13.1
websocket-client==1.8.0
python-multipart==0.0.12
jinja2==3.1.4
aiofiles==24.1.0